code cleanup and remove auth from routes
