+++ packages/react-app/src/constants/Roles.ts
@@ -1,9 +1,9 @@
-export const BANKLESS_ROLE_IDS: Record<string, string> = {
+export const BANKLESS_ROLES = {
 	/**
    * Hardcoded mapping of discord role IDs to some readable names
    * These are in the bankless dao
    */
-	'839005084016312360': 'level-2',
-	'834560235704025108': 'level-1',
-	'911038195397967892': 'bb-core',
+	LEVEL_2: '839005084016312360',
+	LEVEL_1: '834560235704025108',
+	BB_CORE: '911038195397967892',
 };
