+++ packages/react-app/src/hooks/useRoles.ts
@@ -1,7 +1,6 @@
-import { CustomerContext } from '@app/context/CustomerContext';
 import { Role } from '@app/types/Role';
 import { axiosFetcher } from '@app/utils/AxiosUtils';
-import { useContext } from 'react';
+import { useSession } from 'next-auth/react';
 import useSWR from 'swr';
 
 export const useRoles = (): Role[] => {
@@ -9,15 +8,16 @@ export const useRoles = (): Role[] => {
    * This hook is where we can call an API and return the user's roles
 	 * We will probably want to pass in the user at some stage, although currently
 	 * we only use discord roles, so can use the getSession method on the server side
+	 * Note, we currently only support roles in bankless, and need to establish a more complete
+	 * Auth method for multitenancy
    */
-	const { customer } = useContext(CustomerContext);
+	const { data: session } = useSession({ required: false });
 	const { data: roles, error } = useSWR<Role[], unknown>(
-		customer
-			? `/api/auth/roles?customer_id=${customer.customer_id}`
+		session
+			? ['/api/auth/roles', session.accessToken]
 			: null,
-		axiosFetcher
+		axiosFetcher,
 	);
-	console.debug({ roles });
 	if (error) {
 		console.warn(error);
 		return [];
