+++ packages/react-app/src/types/SessionExtended.ts
@@ -1,0 +1,8 @@
+import { Session } from 'next-auth';
+
+/**
+ * Default session object does not include the accessToken as a property
+ */
+export type SessionWithToken = Session & {
+  accessToken: string;
+}
