+++ packages/react-app/src/utils/AxiosUtils.ts
@@ -10,11 +10,24 @@ const axios = $http.create({
 	},
 });
 
+export const axiosTokenFetcher = (url: string, token: string) => axios.get(
+	// fetcher function with token
+	url,
+	{
+		headers: {
+			authorization: `Bearer ${token}`,
+		},
+	}
+).then(res => res.data).catch(error => console.warn(error));
+
+
 export const axiosFetcher = (url: string,) => axios.get(
 	url,
 )
 // strips the messy data.data thing
-	.then(({ data: res }) => res.data);
-// .catch(error => error);
+	.then(({ data: res }) => res.data)
+	.catch(error => {
+		console.warn(error.message ?? error);
+	});
 
 export default axios;
